+++ R/readWorkbook.R
@@ -23,7 +23,7 @@
 #' @param fillMergedCells If TRUE, the value in a merged cell is given to all cells within the merge.
 #' @param skipEmptyCols If `TRUE`, empty columns are skipped.
 #' @param ... additional arguments passed to `wb_to_df()`
-#' @seealso [get_named_regions()] [wb_to_df()]
+#' @seealso [wb_get_named_regions()] [wb_to_df()]
 #' @details Formulae written using write_formula to a Workbook object will not get picked up by read_xlsx().
 #' This is because only the formula is written and left to be evaluated when the file is opened in Excel.
 #' Opening, saving and closing the file with Excel will resolve this.
@@ -57,11 +57,6 @@
 #'   sheet = 2, skipEmptyRows = FALSE,
 #'   cols = c(1, 4), rows = c(1, 3, 4)
 #' )
-#'
-#' ## URL
-#' ##
-#' xlsxFile <- "https://github.com/JanMarvin/openxlsx2/raw/main/inst/extdata/readTest.xlsx"
-#' head(read_xlsx(xlsxFile))
 #' @export
 read_xlsx <- function(
   xlsxFile,
@@ -116,7 +111,7 @@ read_xlsx <- function(
 #' @inheritParams read_xlsx
 #' @details Creates a data.frame of all data in worksheet.
 #' @param ... additional arguments passed to `wb_to_df()`
-#' @seealso [get_named_regions()] [wb_to_df()] [read_xlsx()]
+#' @seealso [wb_get_named_regions()] [wb_to_df()] [read_xlsx()]
 #' @return data.frame
 #' @export
 #' @examples